pyproject.toml
[tool.poetry]
name = "rocket-erp"
version = "0.0.1"
description = ""
authors = ["dimmy2000, quanttyo"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
python-environ = "^0.4.54"
django-environ = "^0.4.5"
djangorestframework = "^3.12.4"
djangorestframework-jwt = "^1.11.0"
Pillow = "^8.2.0"
black = "^21.5b1"
pre-commit = "^2.13.0"
django-bootstrap4 = "^3.0.1"
flake8 = "^3.9.2"
flake8-docstrings = "^1.6.0"
flake8-pytest-style = "^1.4.2"
flake8-codes = "^0.1.1"
flake8-print = "^4.0.0"
django-crum = "^0.7.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
default_section = "THIRDPARTY"
known_first_party = ["api", "accounts"]
known_third_party = []
